+++ src/main/kotlin/introduce/IntroduceMaker.kt
@@ -1,4 +1,4 @@
-package dsl
+package introduce
 
 @DslMarker
 annotation class IntroduceMaker
