+++ src/main/kotlin/introduce/Introduction.kt
@@ -1,3 +1,3 @@
-package dsl
+package introduce
 
 fun introduce(block: PersonBuilder.() -> Unit): Person = PersonBuilder().apply(block).build()
