+++ src/test/kotlin/introduce/IntroductionKtTest.kt
@@ -1,8 +1,8 @@
-package dsl
+package introduce
 
-import dsl.language.Language
-import dsl.skill.Skill
-import dsl.skill.SkillLevel.SOFT
+import introduce.language.Language
+import introduce.skill.Skill
+import introduce.skill.SkillLevel.SOFT
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.collections.shouldHaveSize
 import io.kotest.matchers.shouldBe
